Unskip verify email testcase + improvements
