+++ cypress/integration/2.3-unavailable-product-scenarios.spec.js
@@ -1,22 +1,27 @@
 import availabilityNotifySelectors from '../support/selectors'
-import { loginViaCookies, updateRetry } from '../support/common/support'
+import { loginViaAPI, updateRetry } from '../support/common/support'
 import { testCase1 } from '../support/outputvalidation'
 import {
   generateEmailId,
   updateProductStatus,
+  configureBroadcasterAdapter,
 } from '../support/availability-notify.apis'
 import { subscribeToProductAlerts } from '../support/availability-notify'
 import { MESSAGES } from '../support/utils'
 
 const { name, product, warehouseId, skuId } = testCase1
+const workspace = Cypress.env().workspace.name
+
 const prefix = 'Update product as unavailable'
 
 describe('Updating product as unavailable', () => {
-  loginViaCookies()
+  loginViaAPI()
 
   const email = generateEmailId()
 
   updateProductStatus({ prefix, warehouseId, skuId, unlimited: false })
+
+  configureBroadcasterAdapter(prefix, workspace)
 
   subscribeToProductAlerts({ prefix, product, email, name })
