+++ cypress/support/availability-notify.apis.js
@@ -14,14 +14,14 @@ const version = '1.x'
 export function processAllRequest() {
   it('Process all the requests', updateRetry(3), () => {
     cy.addDelayBetweenRetries(2000)
-    cy.getVtexItems(vtex => {
+    cy.getVtexItems((vtex) => {
       cy.request({
         method: 'GET',
         url: `${getProcessAllRequestAPI()}/${
           vtex.account
         }/${name}/_v/availability-notify/process-all-requests`,
         ...FAIL_ON_STATUS_CODE,
-      }).then(response => {
+      }).then((response) => {
         expect(response.status).to.equal(200)
       })
     })
@@ -36,14 +36,14 @@ export function updateProductStatus({
 }) {
   it(`${prefix} - Update the product status`, updateRetry(3), () => {
     cy.addDelayBetweenRetries(2000)
-    cy.getVtexItems().then(vtex => {
+    cy.getVtexItems().then((vtex) => {
       cy.request({
         method: 'PUT',
         url: updateProductStatusAPI(warehouseId, skuId),
         headers: VTEX_AUTH_HEADER(vtex.apiKey, vtex.apiToken),
         ...FAIL_ON_STATUS_CODE,
         body: { unlimitedQuantity: unlimited, quantity: 0 },
-      }).then(response => {
+      }).then((response) => {
         expect(response.body).to.be.true
       })
     })
@@ -54,13 +54,13 @@ export function notifySearch(prefix) {
   it(`${prefix} - Notify search`, updateRetry(3), () => {
     cy.addDelayBetweenRetries(2000)
 
-    cy.getVtexItems().then(vtex => {
+    cy.getVtexItems().then((vtex) => {
       cy.request({
         method: 'GET',
         url: `https://${vtex.account}.myvtex.com/api/dataentities/notify/search?_schema=reviewsSchema&_fields=email,skuId,name,createdAt`,
         headers: VTEX_AUTH_HEADER(vtex.apiKey, vtex.apiToken),
         ...FAIL_ON_STATUS_CODE,
-      }).then(response => {
+      }).then((response) => {
         expect(response.status).to.equal(200)
       })
     })
@@ -69,7 +69,7 @@ export function notifySearch(prefix) {
 
 export function updateAppSettings(prefix, doShippingSim = false) {
   it(`${prefix} - Configuring app settings`, updateRetry(2), () => {
-    cy.getVtexItems().then(vtex => {
+    cy.getVtexItems().then((vtex) => {
       // Define constants
       const APP_NAME = 'vtex.apps-graphql'
       const APP_VERSION = '3.x'
@@ -106,7 +106,7 @@ export function configureBroadcasterAdapter(prefix, workspace = 'master') {
     `${prefix} - Register target workspace as ${workspace} in ${BROADCASTER_APP}`,
     updateRetry(2),
     () => {
-      cy.getVtexItems().then(vtex => {
+      cy.getVtexItems().then((vtex) => {
         // Define constants
         const APP_NAME = 'vtex.apps-graphql'
         const APP_VERSION = '3.x'
