--- cypress/support/broadcaster.api.js
@@ -1,29 +1,0 @@
-import { FAIL_ON_STATUS_CODE } from './common/constants'
-import { updateRetry } from './common/support'
-
-export function triggerBroadCaster(prefix, skuid) {
-  it(`${prefix} - Triggering broadcaster api`, updateRetry(3), () => {
-    cy.addDelayBetweenRetries(2000)
-    cy.getVtexItems().then(vtex => {
-      cy.getOrderItems().then(order => {
-        cy.request({
-          method: 'POST',
-          url: `http://app.io.vtex.com/vtex.broadcaster/v0/${vtex.account}/${
-            Cypress.env().workspace.name
-          }/notify`,
-          headers: {
-            VtexIdclientAutCookie: order.userAuthCookieValue,
-          },
-          body: {
-            HasStockKeepingUnitModified: true,
-            IdSku: skuid,
-            StockModified: true,
-          },
-          ...FAIL_ON_STATUS_CODE,
-        }).then(response => {
-          expect(response.status).to.equal(200)
-        })
-      })
-    })
-  })
-}
